+++ lib/eventception.rb
@@ -1,5 +1,6 @@
 require 'eventception/event'
 require 'eventception/listener_handler'
 require 'eventception/dispatcher'
+require 'eventception/immutable_dispatcher'
 require 'eventception/base_subscriber'
 require 'eventception/version'