+++ CHANGELOG.md
@@ -3,8 +3,18 @@ All notable changes to this project will be documented in this file. This change
 
 ## [Unreleased]
 
-## 0.0.1 - 2017-07-13
+#### Adds
+
+* Your contribution here.
+* [#13](https://github.com/dcsg/eventception/pull/13): Add an Immutable Dispatcher - [@dcsg](https://github.com/dcsg).
+
+#### Changes
+
+* Your contribution here.
+* [#11](https://github.com/dcsg/eventception/pull/11): Change the license from MIT to GNU LGPLv3 - [@dcsg](https://github.com/dcsg).
+
+## v0.0.1 - 2017-07-13
 
 First release
 
-[Unreleased]: https://github.com/dcsg/eventception/compare/0.0.1...HEAD
+[Unreleased]: https://github.com/dcsg/eventception/compare/v0.0.1...HEAD